requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
colorlog==6.8.2
decorator==5.1.1
distro==1.9.0
Django==5.0.3
django-cors-headers==4.3.1
django-extensions==3.2.3
django-keycloak==0.1.1
django-simple-history==3.5.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
ecdsa==0.18.0
exceptiongroup==1.2.0
executing==2.0.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
ipython==8.22.2
jedi==0.19.1
kombu==5.3.5
matplotlib-inline==0.1.6
minio==7.2.5
openai==1.14.3
parso==0.8.3
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pure-eval==0.2.2
pyasn1==0.6.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-keycloak-client==0.2.3
pytz==2024.1
redis==5.0.3
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
stack-data==0.6.3
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
